requirements.txt
APScheduler==3.6.3
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-APScheduler==1.11.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
ntlm-auth==1.4.0
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pywinrm==0.4.1
requests==2.22.0
requests-ntlm==1.1.0
six==1.13.0
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
xmltodict==0.12.0
